+++ dart/geobase/test/projections_ext/projection_proj4_test.dart
@@ -67,7 +67,7 @@ void main() {
       final inverse = adapter.inverse;
 
       for (var dim = 2; dim <= 4; dim++) {
-        final pointCount = wgs84ToWebMercatorData.length - 1;
+        final pointCount = wgs84ToWebMercatorData.length;
         final source = List.filled(dim * pointCount, 10.0);
         final target = List.filled(dim * pointCount, 10.0);
         for (var i = 0; i < pointCount; i++) {